Add possibility to override the job that handles the stored event
